requirements.txt
sentence-transformers==2.3.1
faiss-cpu==1.7.4
transformers==4.36.2
torch==2.1.1
scikit-learn==1.3.2
PyPDF2==3.0.1
python-docx==1.1.0
wikipedia==1.4.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
nltk==3.8.1
spacy==3.7.2
requests==2.31.0
numpy==1.24.3